name route actions buttons component appropriately
